[MODCON-40]-Added permission and Users clients.
