+++ src/main/java/org/folio/consortia/service/UserTenantService.java
@@ -1,6 +1,5 @@
 package org.folio.consortia.service;
 
-import org.folio.consortia.domain.dto.User;
 import org.folio.consortia.domain.dto.UserEvent;
 import org.folio.consortia.domain.dto.UserTenant;
 import org.folio.consortia.domain.dto.UserTenantCollection;
@@ -115,20 +114,5 @@ public interface UserTenantService {
    */
   void deleteShadowUsers(UUID userId);
 
-  /**
-   * Get existing user.
-   *
-   * @param userId id of user.
-   */
-  User getUser(UUID userId);
 
-  /**
-   * Prepare shadow user from real user.
-   *
-   * @param userId id of user.
-   * @param tenantId id of tenant.
-   *
-   * return user.
-   */
-  User prepareShadowUser(UUID userId, String tenantId);
 }
