+++ src/main/java/org/folio/consortia/service/impl/UserTenantServiceImpl.java
@@ -1,25 +1,21 @@
 package org.folio.consortia.service.impl;
 
-import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.collections4.CollectionUtils;
-import org.folio.consortia.client.UsersClient;
-import org.folio.consortia.domain.dto.Personal;
 import org.folio.consortia.domain.dto.User;
 import org.folio.consortia.domain.dto.UserEvent;
 import org.folio.consortia.domain.dto.UserTenant;
 import org.folio.consortia.domain.dto.UserTenantCollection;
 import org.folio.consortia.domain.entity.TenantEntity;
 import org.folio.consortia.domain.entity.UserTenantEntity;
-import org.folio.consortia.exception.ConsortiumClientException;
 import org.folio.consortia.exception.PrimaryAffiliationException;
 import org.folio.consortia.exception.ResourceNotFoundException;
 import org.folio.consortia.repository.UserTenantRepository;
 import org.folio.consortia.service.ConsortiumService;
 import org.folio.consortia.service.PermissionUserService;
+import org.folio.consortia.service.UserService;
 import org.folio.consortia.service.UserTenantService;
-import org.folio.consortia.utils.HelperUtils;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.FolioModuleMetadata;
 import org.folio.spring.scope.FolioExecutionContextSetter;
@@ -53,13 +49,11 @@ public class UserTenantServiceImpl implements UserTenantService {
   private static final String TENANT_ID = "tenantId";
   private static final Boolean IS_PRIMARY_TRUE = true;
   private static final Boolean IS_PRIMARY_FALSE = false;
-  private static final Integer RANDOM_STRING_COUNT = 5;
-  public static final String PATRON_GROUP = null;
   private final UserTenantRepository userTenantRepository;
   private final FolioExecutionContext folioExecutionContext;
   private final ConversionService converter;
   private final ConsortiumService consortiumService;
-  private final UsersClient usersClient;
+  private final UserService userService;
   private final FolioModuleMetadata folioModuleMetadata;
   private final PermissionUserService permissionUserService;
 
@@ -123,7 +117,7 @@ public class UserTenantServiceImpl implements UserTenantService {
       throw new ResourceNotFoundException(String.format(NOT_FOUND_PRIMARY_AFFILIATION_MSG, USER_ID, userTenantDto.getUserId()));
     }
 
-    User shadowUser = prepareShadowUser(userTenantDto.getUserId(), userTenant.get().getTenant().getId());
+    User shadowUser = userService.prepareShadowUser(userTenantDto.getUserId(), userTenant.get().getTenant().getId());
     createOrUpdateShadowUser(userTenantDto.getUserId(), shadowUser, userTenantDto, currentTenantContext);
 
     try (var context = new FolioExecutionContextSetter(prepareContextForTenant(currentTenantId, folioModuleMetadata, currentTenantContext))) {
@@ -170,7 +164,7 @@ public class UserTenantServiceImpl implements UserTenantService {
     userTenantRepository.deleteByUserIdAndTenantId(userId, tenantId);
 
     try (var context = new FolioExecutionContextSetter(prepareContextForTenant(tenantId, folioModuleMetadata, currentTenantContext))) {
-      User user = getUser(userId);
+      User user = userService.getById(userId);
       deactivateUser(user);
       log.info("User affiliation deleted and user deactivated for user id: {} in the tenant: {}", userId.toString(), tenantId);
     }
@@ -195,38 +189,10 @@ public class UserTenantServiceImpl implements UserTenantService {
     return new UserTenant();
   }
 
-  public User prepareShadowUser(UUID userId, String tenantId) {
-    try (var context = new FolioExecutionContextSetter(prepareContextForTenant(tenantId, folioModuleMetadata, (FolioExecutionContext) folioExecutionContext.getInstance()))) {
-      User user = new User();
-      User userOptional = getUser(userId);
-
-      if (Objects.nonNull(userOptional.getId())) {
-        user.setId(userId.toString());
-        user.setPatronGroup(PATRON_GROUP);
-        user.setUsername(userOptional.getUsername() + HelperUtils.randomString(RANDOM_STRING_COUNT));
-        var userPersonal = userOptional.getPersonal();
-        if (Objects.nonNull(userPersonal)) {
-          Personal personal = new Personal();
-          personal.setLastName(userPersonal.getLastName());
-          personal.setFirstName(userPersonal.getFirstName());
-          personal.setEmail(userPersonal.getEmail());
-          personal.setPreferredContactTypeId(userPersonal.getPreferredContactTypeId());
-          user.setPersonal(personal);
-        }
-        user.setPatronGroup(userOptional.getPatronGroup());
-        user.setActive(true);
-      } else {
-        log.warn("Could not find real user with id: {} in his home tenant: {}", userId.toString(), tenantId);
-        throw new ResourceNotFoundException(USER_ID, userId.toString());
-      }
-      return user;
-    }
-  }
-
   private void createOrUpdateShadowUser(UUID userId, User shadowUser, UserTenant userTenantDto, FolioExecutionContext folioExecutionContext) {
     log.info("Going to create or update shadow user with id: {} in the desired tenant: {}", userId.toString(), userTenantDto.getTenantId());
     try (var context = new FolioExecutionContextSetter(prepareContextForTenant(userTenantDto.getTenantId(), folioModuleMetadata, folioExecutionContext))) {
-      User user = getUser(userId);
+      User user = userService.getById(userId);
       if (Objects.nonNull(user.getActive())) {
         activateUser(user);
       } else {
@@ -235,32 +201,11 @@ public class UserTenantServiceImpl implements UserTenantService {
     }
   }
 
-  /**
-   * Gets user by id.
-   * <p>
-   * This method will be called to get User from MOD_USERS module based on the tenant schema present in the folioContext.
-   *
-   * @param userId user id.
-   */
-  public User getUser(UUID userId) {
-    try {
-      log.info("Getting user by userId {}.", userId);
-      return usersClient.getUsersByUserId(String.valueOf(userId));
-    } catch (FeignException.NotFound e) {
-      log.info("User with userId {} does not exist in schema, going to use new one", userId);
-      return new User();
-    } catch (FeignException.Forbidden e) {
-      throw new ConsortiumClientException(e);
-    } catch (FeignException e) {
-      throw new IllegalStateException(e);
-    }
-  }
-
   private void createActiveUserWithPermissions(User user) {
     log.info("Creating permissionUser for userId {} with empty set of permissions", user.getId());
-    permissionUserService.createWithPermissions(UUID.randomUUID().toString(), user.getId(), List.of());
+    permissionUserService.createWithEmptyPermissions(UUID.randomUUID().toString(), user.getId(), List.of());
     log.info("Creating user with id {}.", user.getId());
-    usersClient.saveUser(user);
+    userService.createUser(user);
   }
 
   private void activateUser(User user) {
@@ -269,7 +214,7 @@ public class UserTenantServiceImpl implements UserTenantService {
     } else {
       user.setActive(true);
       log.info("Updating User with id '{}' with active 'true'. ", user.getId());
-      usersClient.updateUser(user.getId(), user);
+      userService.updateUser(user.getId(), user);
     }
   }
 
@@ -279,7 +224,7 @@ public class UserTenantServiceImpl implements UserTenantService {
     } else {
       user.setActive(false);
       log.info("Updating User with id '{}' with active 'false'. ", user.getId());
-      usersClient.updateUser(user.getId(), user);
+      userService.updateUser(user.getId(), user);
     }
   }
 
@@ -291,7 +236,7 @@ public class UserTenantServiceImpl implements UserTenantService {
       log.info("Removing orphaned shadow users from all tenants exist in consortia for the user: {}", userId);
       tenantIds.forEach(tenantId -> {
         try (var context = new FolioExecutionContextSetter(prepareContextForTenant(tenantId, folioModuleMetadata, folioExecutionContext))) {
-          usersClient.deleteUser(userId.toString());
+          userService.deleteById(userId.toString());
           log.info("Removed shadow user: {} from tenant : {}", userId, tenantId);
         }
       });
