+++ src/test/java/org/folio/consortia/service/UserTenantServiceTest.java
@@ -1,6 +1,5 @@
 package org.folio.consortia.service;
 
-import feign.FeignException;
 import org.folio.consortia.client.PermissionsClient;
 import org.folio.consortia.client.UsersClient;
 import org.folio.consortia.domain.converter.UserTenantConverter;
@@ -84,6 +83,8 @@ class UserTenantServiceTest {
   private FolioModuleMetadata folioModuleMetadata;
   @Mock
   private PermissionUserService permissionUserService;
+  @Mock
+  private UserService userService;
 
   /* Success cases */
   @Test
@@ -207,7 +208,8 @@ class UserTenantServiceTest {
     when(consortiumRepository.findById(UUID.fromString(CONSORTIUM_ID)))
       .thenReturn(Optional.of(createConsortiumEntity()));
     when(userTenantRepository.findByUserIdAndIsPrimary(any(), any())).thenReturn(Optional.of(userTenant));
-    when(usersClient.getUsersByUserId(any())).thenReturn(createUserEntity(false));
+    when(userService.getById(any())).thenReturn(createUserEntity(false));
+    when(userService.prepareShadowUser(any(),any())).thenReturn(createUserEntity(false));
     doNothing().when(usersClient).updateUser(any(), any(User.class));
     when(userTenantRepository.save(userTenant)).thenReturn(userTenant);
     when(folioExecutionContext.getTenantId()).thenReturn("diku");
@@ -231,7 +233,8 @@ class UserTenantServiceTest {
     when(consortiumRepository.findById(UUID.fromString(CONSORTIUM_ID)))
       .thenReturn(Optional.of(createConsortiumEntity()));
     when(userTenantRepository.findByUserIdAndIsPrimary(any(), any())).thenReturn(Optional.of(userTenant));
-    when(usersClient.getUsersByUserId(any())).thenReturn(createUserEntity(true));
+    when(userService.getById(any())).thenReturn(createUserEntity(true));
+    when(userService.prepareShadowUser(any(),any())).thenReturn(createUserEntity(true));
     doNothing().when(usersClient).updateUser(any(), any(User.class));
     when(userTenantRepository.save(userTenant)).thenReturn(userTenant);
     when(folioExecutionContext.getTenantId()).thenReturn("diku");
@@ -257,7 +260,8 @@ class UserTenantServiceTest {
     when(consortiumRepository.findById(UUID.fromString(CONSORTIUM_ID)))
       .thenReturn(Optional.of(createConsortiumEntity()));
     when(userTenantRepository.findByUserIdAndIsPrimary(any(), any())).thenReturn(Optional.of(userTenant));
-    when(usersClient.getUsersByUserId(any())).thenReturn(createNullUserEntity());
+    when(userService.getById(any())).thenReturn(createNullUserEntity());
+    when(userService.prepareShadowUser(any(),any())).thenReturn(createNullUserEntity());
     when(permissionsClient.get(any())).thenReturn(permissionUserCollection);
     doNothing().when(usersClient).saveUser(any());
     when(userTenantRepository.save(userTenant)).thenReturn(userTenant);
@@ -284,7 +288,8 @@ class UserTenantServiceTest {
     when(consortiumRepository.findById(UUID.fromString(CONSORTIUM_ID)))
       .thenReturn(Optional.of(createConsortiumEntity()));
     when(userTenantRepository.findByUserIdAndIsPrimary(any(), any())).thenReturn(Optional.of(userTenant));
-    when(usersClient.getUsersByUserId(any())).thenReturn(createNullUserEntity());
+    when(userService.getById(any())).thenReturn(createNullUserEntity());
+    when(userService.prepareShadowUser(any(),any())).thenReturn(createNullUserEntity());
     when(permissionsClient.get(any())).thenReturn(permissionUserCollection);
     when(permissionsClient.create(any())).thenReturn(permissionUser);
     doNothing().when(usersClient).saveUser(any());
@@ -329,7 +334,8 @@ class UserTenantServiceTest {
     userTenant.setIsPrimary(false);
 
     when(consortiumRepository.existsById(UUID.fromString(CONSORTIUM_ID))).thenReturn(true);
-    when(usersClient.getUsersByUserId(any())).thenReturn(createNullUserEntity());
+    when(userService.getById(any())).thenReturn(createNullUserEntity());
+    when(userService.prepareShadowUser(any(),any())).thenReturn(createNullUserEntity());
     when(userTenantRepository.findByUserIdAndTenantId(userId, tenantId))
       .thenReturn(Optional.of(userTenant));
     doNothing().when(userTenantRepository).deleteByUserIdAndTenantId(userId, tenantId);
@@ -448,7 +454,7 @@ class UserTenantServiceTest {
     when(consortiumRepository.findById(UUID.fromString(CONSORTIUM_ID)))
       .thenReturn(Optional.of(createConsortiumEntity()));
     when(userTenantRepository.findByUserIdAndIsPrimary(any(), any())).thenReturn(Optional.of(userTenant));
-    when(usersClient.getUsersByUserId(any())).thenThrow(FeignException.Forbidden.class);
+    when(userService.getById(any())).thenThrow(org.folio.consortia.exception.ConsortiumClientException.class);
     doNothing().when(usersClient).updateUser(any(), any(User.class));
     when(userTenantRepository.save(userTenant)).thenReturn(userTenant);
     when(folioExecutionContext.getTenantId()).thenReturn("diku");
@@ -475,7 +481,7 @@ class UserTenantServiceTest {
     when(consortiumRepository.findById(UUID.fromString(CONSORTIUM_ID)))
       .thenReturn(Optional.of(createConsortiumEntity()));
     when(userTenantRepository.findByUserIdAndIsPrimary(any(), any())).thenReturn(Optional.of(userTenant));
-    when(usersClient.getUsersByUserId(any())).thenThrow(FeignException.NotFound.class);
+    when(userService.getById(any())).thenThrow(org.folio.consortia.exception.ResourceNotFoundException.class);
     doNothing().when(usersClient).updateUser(any(), any(User.class));
     when(userTenantRepository.save(userTenant)).thenReturn(userTenant);
     when(folioExecutionContext.getTenantId()).thenReturn("diku");
@@ -501,7 +507,7 @@ class UserTenantServiceTest {
     when(consortiumRepository.findById(UUID.fromString(CONSORTIUM_ID)))
       .thenReturn(Optional.of(createConsortiumEntity()));
     when(userTenantRepository.findByUserIdAndIsPrimary(any(), any())).thenReturn(Optional.of(userTenant));
-    when(usersClient.getUsersByUserId(any())).thenThrow(FeignException.class);
+    when(userService.getById(any())).thenThrow(java.lang.IllegalStateException.class);
     when(folioExecutionContext.getTenantId()).thenReturn("diku");
     when(folioExecutionContext.getInstance()).thenReturn(folioExecutionContext);
     Map<String, Collection<String>> okapiHeaders = new HashMap<>();
